+++ modules/hanna-utils/src/_/printDate.ts
@@ -48,9 +48,13 @@ const monthsByLang: OpenRecord<'is' | 'en' | 'pl', Array<string>> = {
 };
 
 /**
- * Very simple, very stupid, standalone date formatter for Icelandic, English and Polish.
+ * Very simple, very stupid, standalone date formatter for Icelandic (default),
+ * English and Polish.
  *
- * Just prints the full date (day, month, year).
+ * Just prints the full date (day, month, year). No bells. No whistles.
+ * No options.
+ *
+ * @see https://www.npmjs.com/package/@reykjavik/hanna-utils#printdate
  */
 export const printDate = (date: string | Date, lang?: string): string => {
   date = typeof date === 'string' ? new Date(date) : date;